Hardcoding comic ID, making content scrollable
Lots of the comics from the API do not exist, so rather than picking a random number and potentially getting the error state I'm just picking a specific ID for the sake of demonstration.

I added the verticalScroll modifier to the main Column because the content can overflow the screen.